Extract ReflectionInfoExtractor
